requirements.txt
pyrogram
TgCrypto
psycopg2-binary
